WTA #71: Added endpoints and mappers for getTotalTimeForProject.
